src/app/features/protected/user-info/user-info.component.ts
import { Component } from '@angular/core';
import { AuthService } from '../../../core/services/auth.service';
import { CommonModule } from '@angular/common';

@Component({
  selector: 'app-user-info',
  standalone: true,
  imports: [CommonModule],
  templateUrl: './user-info.component.html',
  styleUrl: './user-info.component.css'
})
export class UserInfoComponent {
  user: { name: string; email: string } = { name: '', email: '' };
  constructor(private authService: AuthService){
    this.user = this.authService.getUserInfo();
  }

}
